pyproject.toml
[build-system]
requires = ["setuptools>=69.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flowfetch"
version = "0.1.0"
description = "A Linux-first terminal downloader with progress display and optional safe extraction."
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
authors = [
  { name = "Juntao He", email = "hejuntao@bupt.edu.cn" },
]
dependencies = [
  "httpx>=0.27,<1.0",
  "rich>=13.7,<14.0",
]
keywords = ["cli", "download", "linux", "terminal", "archive"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Environment :: Console",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: POSIX :: Linux",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Internet :: WWW/HTTP",
  "Topic :: Utilities",
]

[project.urls]
Homepage = "https://github.com/MRT-8/FlowFetch"
Repository = "https://github.com/MRT-8/FlowFetch"
Issues = "https://github.com/MRT-8/FlowFetch/issues"

[project.scripts]
flowfetch = "downloader:main"

[project.optional-dependencies]
media = [
  "pillow>=11,<12",
]
release = [
  "pyinstaller>=6,<7",
]

[tool.setuptools]
py-modules = ["downloader"]
